Introduce user-only configuration. #238
That commit allows to decouple what the user can customize (which is mostly optional), and we want really
want to be there (which is requiring almost everything to be provided). That requires being able to merge
user configurations with regular configurations, and to create a regular configuration from a user
configuration (for instance, when adding a new language config via the user configuration only), which can
fail if some options are not provided.